Feat/edit sale

+++ src/main/java/fr/eni/projet/eniencheres/dal/utilisateur/UtilisateurRowMapper.java
@@ -17,7 +17,7 @@ public class UtilisateurRowMapper implements RowMapper<Utilisateur> {
         utilisateur.setPrenom(rs.getString("prenom"));
         utilisateur.setEmail(rs.getString("email"));
         utilisateur.setTelephone(rs.getString("telephone"));
-        utilisateur.setCredit(rs.getDouble("credit"));
+        utilisateur.setCredit(rs.getInt("credit"));
 
         AdresseRowMapper adresseRowMapper = new AdresseRowMapper();
         Adresse adresse = adresseRowMapper.mapRow(rs, rowNum);
